requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# PDF Processing & OCR
PyPDF2==3.0.1
pdfplumber==0.10.3
PyMuPDF==1.23.8
pytesseract==0.3.10
pdf2image==1.17.0
pillow==10.2.0

# Text Processing
python-docx==1.1.0
beautifulsoup4==4.12.3
lxml==5.1.0
unstructured==0.12.0

# Embeddings & Vector Search
sentence-transformers==2.3.1
faiss-cpu==1.7.4
transformers==4.36.2
torch==2.1.2

# Search & Retrieval
rank-bm25==0.2.2

# ML & NLP
numpy==1.26.3
scikit-learn==1.4.0
spacy==3.7.2

# Data Processing
pandas==2.1.4
pyyaml==6.0.1

# Async & Task Queue
celery==5.3.6
redis==5.0.1

# HTTP & API
httpx==0.26.0
requests==2.31.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
python-dotenv==1.0.1
black==24.1.1
ruff==0.1.14

# Logging & Monitoring
loguru==0.7.2